Rename `bids_units.py` in `bids.py`
